changed: tests split into modules
