+++ tests/test_dbscan.py
@@ -1,0 +1,182 @@
+import unittest
+
+import numpy as np
+from sklearn.datasets import make_blobs
+from sklearn.datasets import make_circles
+from sklearn.datasets import make_moons
+from sklearn.preprocessing import StandardScaler
+
+from dislib.cluster import DBSCAN
+from dislib.data import load_data
+
+
+class DBSCANTest(unittest.TestCase):
+    def test_n_clusters_blobs(self):
+        n_samples = 1500
+
+        # Test blobs
+        x, y = make_blobs(n_samples=n_samples, n_features=2, random_state=8)
+        dbscan = DBSCAN(grid_dim=1, eps=.3)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 3)
+
+    def test_n_clusters_circles(self):
+        n_samples = 1500
+
+        # Test circles
+        x, y = make_circles(n_samples=n_samples, factor=.5, noise=.05)
+        dbscan = DBSCAN(grid_dim=1, eps=.15)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_moons(self):
+        n_samples = 1500
+
+        # Test moons
+        x, y = make_moons(n_samples=n_samples, noise=.05)
+        dbscan = DBSCAN(grid_dim=1, eps=.3)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_aniso(self):
+        n_samples = 1500
+
+        # Test aniso
+        x, y = make_blobs(n_samples=n_samples, random_state=170)
+        dbscan = DBSCAN(grid_dim=1, eps=.15)
+        transformation = [[0.6, -0.6], [-0.4, 0.8]]
+        x = np.dot(x, transformation)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 4)
+
+    def test_n_clusters_blobs_max_samples(self):
+        n_samples = 1500
+
+        # Test blobs
+        x, y = make_blobs(n_samples=n_samples, n_features=2, random_state=8)
+        dbscan = DBSCAN(grid_dim=1, eps=.3, max_samples=500)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 3)
+
+    def test_n_clusters_circles_max_samples(self):
+        n_samples = 1500
+
+        # Test circles
+        x, y = make_circles(n_samples=n_samples, factor=.5, noise=.05)
+        dbscan = DBSCAN(grid_dim=1, eps=.15, max_samples=500)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_moons_max_samples(self):
+        n_samples = 1500
+
+        # Test moons
+        x, y = make_moons(n_samples=n_samples, noise=.05)
+        dbscan = DBSCAN(grid_dim=1, eps=.3, max_samples=500)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_aniso_max_samples(self):
+        n_samples = 1500
+
+        # Test aniso
+        x, y = make_blobs(n_samples=n_samples, random_state=170)
+        dbscan = DBSCAN(grid_dim=1, eps=.15, max_samples=500)
+        transformation = [[0.6, -0.6], [-0.4, 0.8]]
+        x = np.dot(x, transformation)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 4)
+
+    def test_n_clusters_blobs_grid(self):
+        n_samples = 1500
+
+        # Test blobs
+        x, y = make_blobs(n_samples=n_samples, n_features=2, random_state=8)
+        dbscan = DBSCAN(grid_dim=2, eps=.3, max_samples=300)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 3)
+
+    def test_n_clusters_circles_grid(self):
+        n_samples = 1500
+
+        # Test circles
+        x, y = make_circles(n_samples=n_samples, factor=.5, noise=.05)
+        dbscan = DBSCAN(grid_dim=2, eps=.15, max_samples=700)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_moons_grid(self):
+        n_samples = 1500
+
+        # Test moons
+        x, y = make_moons(n_samples=n_samples, noise=.05)
+        dbscan = DBSCAN(grid_dim=2, eps=.3, max_samples=600)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 2)
+
+    def test_n_clusters_aniso_grid(self):
+        n_samples = 1500
+
+        # Test aniso
+        x, y = make_blobs(n_samples=n_samples, random_state=170)
+        dbscan = DBSCAN(grid_dim=2, eps=.15, max_samples=500)
+        transformation = [[0.6, -0.6], [-0.4, 0.8]]
+        x = np.dot(x, transformation)
+        x = StandardScaler().fit_transform(x)
+        dataset = load_data(x=x, y=y, subset_size=300)
+        dbscan.fit(dataset)
+        n_clusters = np.unique(dbscan.labels_)
+        n_clusters = n_clusters[n_clusters >= 0].size
+        self.assertEqual(n_clusters, 4)
+
+
+def main():
+    unittest.main()
+
+
+if __name__ == '__main__':
+    main()
